+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.11.0
+
+- DONE: Link hte new modal of pictures to the waypoint dashboard
+- FIX: `name`and `address`property for `ScreenWaypointGalery`
+- ADD: A new look for disabled `CButton`
+- ADD: The photo button in `CWaypointButtons`now contain the number of pictures and is disabled if no pictures
+- ADD: `waypointPictureCollection`to the waypoint state
+- ADD: newly created modal to the waypoint dashboard
+
 # 0.10.0
 
 - DONE: Picture gallery of the waypoint